requirements.txt
# Core ML
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# Transformers & NLP
transformers>=4.40.0
tokenizers>=0.19.0
# sentencepiece>=0.1.99

# Audio
librosa>=0.10.1
soundfile>=0.12.1

# Image/Video processing
opencv-python-headless>=4.9.0.80
Pillow>=10.2.0

# Utilities & metrics
scikit-learn>=1.4.0
tqdm>=4.66.0
pyyaml>=6.0.1
numpy>=1.26.0

# Optional explainability & trees
shap>=0.44.1
lightgbm>=4.3.0